Add order_ids to adjustments where they were missing
* Adds a migration to add order_ids to adjustments to pass new
validations.
* Change importer/order to use to_d instead of to_f
* Update specs

+++ core/lib/spree/core/importer/order.rb
@@ -88,7 +88,7 @@ module Spree
             begin
               adjustment = order.adjustments.build(
                 order:  order,
-                amount: a[:amount].to_f,
+                amount: a[:amount].to_d,
                 label:  a[:label]
               )
               adjustment.save!
@@ -105,7 +105,7 @@ module Spree
             begin
               payment, success = order.contents.add_payment(
                 payment_params: {
-                  amount: p[:amount].to_f,
+                  amount: p[:amount].to_d,
                   state: p.fetch(:state, 'completed'),
                   payment_method: Spree::PaymentMethod.find_by_name!(p[:payment_method])
                 }
